+++ tests/replica/020_reference.t.ts
@@ -54,6 +54,11 @@ StartTest(t => {
         replica1.removeEntity(tomSoyer)
 
         t.isDeeply(markTwain.books, new Set(), 'Correctly resolved reference')
+
+        //--------------------
+        replica1.addEntity(tomSoyer)
+
+        t.isDeeply(markTwain.books, new Set([ tomSoyer ]), 'Correctly resolved reference #2')
     })
 
 
@@ -112,6 +117,13 @@ StartTest(t => {
         replica1.commit()
 
         t.isDeeply(markTwain.books, new Set(), 'Correctly resolved reference')
+
+        //--------------------
+        replica1.addEntity(tomSoyer)
+
+        replica1.commit()
+
+        t.isDeeply(markTwain.books, new Set([ tomSoyer ]), 'Correctly resolved reference')
     })