+++ CHANGELOG.md
@@ -14,7 +14,10 @@
 
 - Fixed a bug, that on-demand calculations for lazy identifiers were performed in separate
 transaction. This was breaking the `ProposedOrPrevious` effect. Such calculations are now
-performed in the currently running transaction
+performed in the currently running transaction.
+
+- Fixed a bug, that removing and adding an entity back to the graph right away was not updating
+the related buckets.
 
 ## 1.0.1        2020-04-07 19:56